Move Stage out of boot_animation.py
