+++ src/animations/start_computer_bios.py
@@ -24,8 +24,10 @@ This is the boot animation from the first computer
 # init colors
 import curses
 
-from src.core.boot_animation.boot_animation import StyledText, Stage, BootAnimation
+from src.core.boot_animation.boot_animation import BootAnimation
+from src.core.boot_animation.stage import Stage
 from src.core.boot_animation.step import Step
+from src.core.boot_animation.styled_text import StyledText
 from src.core.render import CursesRenderer
